@@ -1,7 +1,7 @@
-require('normalize.css/normalize.css')
-require('../css/main.css')
-require('./page.css')
-const img = require('../img/unicorn.jpg')
+import '../scss/main.scss'
+import './page.scss'
+import img from '../img/unicorn.jpg'
+
 
 document.addEventListener('DOMContentLoaded', () => {
   console.log('DOMContentLoaded', 'page-about')
