@@ -1,8 +1,8 @@
 import { Bar } from '../js/bar.es6'
 const { Foo } = require('../js/foo.cjs')
-require('normalize.css/normalize.css')
-require('../css/main.css')
-require('./page.css')
+// import 'normalize.css/normalize.css'
+import '../scss/main.scss'
+import './page.scss'
 
 document.addEventListener('DOMContentLoaded', () => {
   console.log('DOMContentLoaded', 'page-index')
